Create unfold
unfold returns a stream by constantly applying the function f to the seed value s of type 'a, either returning NONE when the stream is finished, or an element of a stream along with a new seed value, gradually building up a stream. For example you could start with the seed value as 1 and have f(s) => s + 1, effectively creating a stream of all the integers starting with 1 and counting up.