Create nth, take, and drop
"nth" takes a stream and an int n and returns the nth element of that stream. "take" takes a stream and an int n  and outputs the stream up to the nth element. "drop" takes a stream and an int n and returns the stream from the n+1nth element onwards.